Login and SignUp
the logic and html for the sign in and attempted sign up feature connecting to a mongoDB. Need to fix up the sign up feature. Login feature works as it check in the database so we can just initialize if we want. Have not connected it to rest of of the project but should be a simple call to the rest of the frontend